Behaviour Tree classes restructured
The Tree class is no longer in use, it's functionality has been moved into the individual animal's class. The behaviour tree is the same for all animals so this has been moved into the AnimalAI class. This meant that the function to build the tree needed to be parameterised to tune behaviour for each animal. This is why the building of the tree was moved into the individual animal's code.