Added Travel Type and Starting Player Prefs
Travel type for each animal needs to be known for how the animal will move in the game world. An enum has been created for this, including land, water and air.

Player prefs will be needed to save data between gameplay sessions. This is being incorporated but may be delayed to ensure main mechanics are finished first.