Accessors in AnimalAI & more LeafNode Logic
FollowPlayer - If the player is near enough to the player, the agent's nav mesh destination is removed and the node succeeds as they have successfully followed the player. Otherwise, every 3 seconds, the nav agent's destination is set to the player's current position, and the node still succeeds.

IsPlayerNear - Adjusted so logic for animal looking at player is no longer here (see WaitForPlayer). IsThreated is set to false and node fails if player is not near, succeeds if it is.

FightOrFlight - switch statement to have different behaviour for different aggression levels. Succeeds as long as an aggression level is found. Only calculates new destination if navAgent.remainingDistance < 4f.
*** NOTE *** - Large frame drop when calculating destination sometimes. Unknown reason.

WaitForPlayer - Logic for looking at player when in close proximity and tamed has been moved here

Wander - logic adjusted so path is calculated before the agent reaches a destination, resulting in smoother movement.