Medium Aggression Fight Or Flight Behaviour
Fight or flight behaviour has been added for animals with medium levels of aggression. They do not attack instantly, but after being threatened twice, or if 2 seconds after being threatened the player has not left the animal's danger distance, it will run towards the player to attack. The attacking behaviour itself has not been added yet.

Animals now have a separate DangerDist which is half that of the DetectDist. If in the detect dist, a wild animal will stare at the player. If in the danger dist, currently the animal will be threatened, but this logic will be tied to either being attacked, or a player running nearby, at a later date.